Now exits fullscreen video if the video is blocked
Basically just added wistiaEmbed.cancelFullscreen(); in the block() function. There was and are a few functions that are called that theoretically exit the full screen but I don't think that they actually apply here with a fullscreen flash video. I left them in for good measure. So if you are in full screen and access is blocked, wistia will pause and the login thing will show up.

Changes tested on Chrome and Firefox in Ubuntu on a dev server, and it seemed to fix the bug completely, hopefully ths is sufficent everywhere.